The DHT22 Humidity And Temperature Sensor
The DHT22 sensor is a digital sensor that can measure temperature and humidity with higher accuracy . It uses a capacitive humidity sensor and a thermistor to measure the surrounding air and converts the readings to digital signals.

When used with Arduino, the DHT22 sensor is interfaced , using a single data wire. The Arduino reads the sensor data and converts it to temperature and humidity values, which can then be used for various applications such as climate monitoring, weather stations, and home automation systems.